+++ tools/remote_execution/blast/src/re_cli/core/script_builder/__init__.py
@@ -2,16 +2,12 @@
 
 from .builder import (
     create_bootstrap,
-    GitCloneConfig,
-    GitCloneMethod,
     RunnerConfig,
     RunnerScriptBuilder,
 )
 
 
 __all__ = [
-    "GitCloneConfig",
-    "GitCloneMethod",
     "RunnerScriptBuilder",
     "RunnerConfig",
     "create_bootstrap",
