+++ tools/remote_execution/blast/src/re_cli/core/script_builder/templates/exit.sh
@@ -1,0 +1,4 @@
+# MODULE: Exit
+# Final module - exits with the user script's exit code
+
+exit $SCRIPT_EXIT_CODE
