+++ tools/remote_execution/blast/src/re_cli/core/script_builder/templates/git_templates/git_clone.sh
@@ -1,20 +1,12 @@
 # MODULE: Git Clone
-# Handles git repo cloning with cache support
-
-REPO_DIR=""
+# Shallow clone of git repo
 
 if [[ -n "$GIT_REPO" ]]; then
-    # Check if REPO_CACHE is set (EFS/daemonset provided cache)
-    if [[ -n "$REPO_CACHE" && -d "$REPO_CACHE/.git" ]]; then
-        echo "[Runner] Using repo cache from $REPO_CACHE"
-        cp -r "$REPO_CACHE" repo
-        cd repo
-        REPO_DIR="$(pwd)"
-        export REPO_DIR
-    else
-{{git_clone_method}}
-    fi
-
+    echo "[Runner] Cloning $GIT_REPO (shallow)..."
+    git clone --depth=1 "$GIT_REPO" repo
+    cd repo
+    REPO_DIR="$(pwd)"
+    export REPO_DIR
     echo "[Runner] REPO_DIR=$REPO_DIR"
 else
     echo "[Runner] No git repo specified, skipping clone"
