+++ tools/remote_execution/blast/src/re_cli/core/script_builder/templates/upload_outputs.sh
@@ -34,5 +34,3 @@ if [[ -d "$LOG_DIR" ]]; then
 else
     echo "[Runner] No logs to upload (LOG_DIR does not exist)"
 fi
-
-exit $SCRIPT_EXIT_CODE
